feat(core): move ecl parsing and sequencing services to core module
Register them in their own EclPlugin.
