+++ core/com.b2international.snowowl.core/META-INF/MANIFEST.MF
@@ -53,6 +53,7 @@ Export-Package: com.auth0.jwt,
  com.b2international.snowowl.core.context,
  com.b2international.snowowl.core.date,
  com.b2international.snowowl.core.domain,
+ com.b2international.snowowl.core.ecl,
  com.b2international.snowowl.core.emf,
  com.b2international.snowowl.core.events,
  com.b2international.snowowl.core.events.bulk,
