+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/ecl/EclPlugin.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2022 B2i Healthcare Pte Ltd, http://b2i.sg
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.b2international.snowowl.core.ecl;
+
+import org.eclipse.xtext.parser.IParser;
+import org.eclipse.xtext.serializer.ISerializer;
+import org.eclipse.xtext.validation.IResourceValidator;
+
+import com.b2international.snomed.ecl.EclStandaloneSetup;
+import com.b2international.snowowl.core.config.SnowOwlConfiguration;
+import com.b2international.snowowl.core.plugin.Component;
+import com.b2international.snowowl.core.setup.Environment;
+import com.b2international.snowowl.core.setup.Plugin;
+import com.google.inject.Injector;
+
+/**
+ * @since 8.2.0
+ */
+@Component
+public class EclPlugin extends Plugin {
+
+	@Override
+	public void init(SnowOwlConfiguration configuration, Environment env) throws Exception {
+		final Injector injector = new EclStandaloneSetup().createInjectorAndDoEMFRegistration();
+		env.services().registerService(EclParser.class, new DefaultEclParser(injector.getInstance(IParser.class), injector.getInstance(IResourceValidator.class)));
+		env.services().registerService(EclSerializer.class, new DefaultEclSerializer(injector.getInstance(ISerializer.class)));
+	}
+	
+}
